Fix iperf_accept in server mode
In udp control sock mode, iperf_accept is called whenever select comes
out as we don't have a dedicated socket for listening to events. In case
of udp control socket, both listener and control socket are one and the
same. To avoid mishandling events, add additional checks such that it
won't land in iperf_accept handler